# feat: use RcppParallel and pure C++ distance functions