--- a/.Rbuildignore
+++ b/.Rbuildignore
@@ -12,3 +12,8 @@ paper.bib
 docs
 LICENSE
 ^_pkgdown.yml$
+^\.gitignore$
+^inst/include/\.DS_Store$
+^\.\.Rcheck$
+^\.git$
+^\.vscode$