--- a/R/utils.R
+++ b/R/utils.R
@@ -19,6 +19,7 @@
 
 #' @useDynLib philentropy, .registration = TRUE
 #' @importFrom Rcpp sourceCpp
+#' @importFrom RcppParallel RcppParallelLibs
 NULL
 
 .onUnload <- function(libpath) {