--- a/man/dist_one_one.Rd
+++ b/man/dist_one_one.Rd
@@ -8,7 +8,7 @@ dist_one_one(
   P,
   Q,
   method,
-  p = NA_real_,
+  p = NULL,
   testNA = TRUE,
   unit = "log",
   epsilon = 1e-05