… NAMESPACE
… R/distance.R
… src/dist_matrix.cpp
… src/distances.h
… src/distances_internal.h
… tests/testthat/test-distances.R
… vignettes/Many_Distances.html