Hello @HajkD & @Nowosad!

Thank you for the wonderful philentropy package. I have been using it a good bit lately under some more computationally demanding cases and was curious to see whether I could optimize the distance calculations further. Thanks in advance for your consideration.

In this PR the core C++ implementations for distance calculations have been heavily optimized to make use of RcppParallel. 
These patterns of optimization significantly reduce computation time and memory allocation, making the package much more efficient for large-scale analyses. 

The main change is implementing the calculations in pure C++ with templates so that we could easily re-use the core logic for each of the distance methods (which you have done such a great job pulling together in this package) both in the RcppParallel workers as well as the one-to-one user facing functions. I have also simplified some of the R wrapper code. 

Benchmarks on my machine show a speedup of over 100x for `distance()` and over 70x for `dist_many_many()` on key methods like euclidean, minkowski, and jensen-shannon. 

I first implemented parallel versions of functions for side by side testing, then incrementally replaced existing code with the new backend and squashed my final result into one commit. Then I updated the benchmarks in the Many Distances vignette.

Throughout I relied heavily on the existing unit tests to make sure that my changes were sound. However, in a couple cases (K divergence, Topsoe) I found test cases that appeared, to me, to be incorrect. Cases involving 0*log(0) would be evaluated as NaN which would propagate to the final result because of R's standard rules. However, conventionally in information theory this expression would generally be interpreted as zero (rather than undefined). 

I would be happy to roll back any changes if you feel that it would be a problem to make them, or implement parallel functionality through an additional argument or option for backward compatibility. I also would be willing to add more tests. Finally understand if you would prefer not incorporate this PR at all--but perhaps you will find the patterns thought provoking or useful in future updates!

<details>

## Before

``` r
# install.packages("bench")
library(philentropy)
library(bench)

set.seed(42)

n_dist <- 300 # Number of distributions
n_bins <- 200  # Number of bins per distribution

dist_matrix_rows <- matrix(runif(n_dist * n_bins), nrow = n_dist)

p <- 2
unit <- "log2"

cat("Benchmarking distance matrix functions with", n_dist, "distributions of", n_bins, "bins each.\n\n")
#> Benchmarking distance matrix functions with 300 distributions of 200 bins each.

cat("1. Benchmark: 'euclidean' distance\n")
#> 1. Benchmark: 'euclidean' distance
euclidean_bench <- bench::mark(
  `distance(method = 'euclidean')` = distance(dist_matrix_rows, method = "euclidean", test.na = FALSE),
  `dist_many_many(method = 'euclidean')` = dist_many_many(dist_matrix_rows, dist_matrix_rows, method = "euclidean", testNA = FALSE),
  min_iterations = 3,  max_iterations = 20,
  check = \(a, b) all(a == b)
)
#> Metric: 'euclidean'; comparing: 300 vectors.
#> Metric: 'euclidean'; comparing: 300 vectors.
#> Metric: 'euclidean'; comparing: 300 vectors.
#> Metric: 'euclidean'; comparing: 300 vectors.
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
print(euclidean_bench)
#> # A tibble: 2 × 13
#>   expression      min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time
#>   <bch:expr>    <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm>
#> 1 distance(met… 208ms  242ms      3.07     217MB     7.17     3     7      976ms
#> 2 dist_many_ma… 216ms  216ms      4.58     284MB    15.3      3    10      655ms
#> # ℹ 4 more variables: result <list>, memory <list>, time <list>, gc <list>
plot(euclidean_bench)
```

![](https://i.imgur.com/M68k8hd.png)<!-- -->

``` r

cat("\n2. Benchmark: 'minkowski' distance (p=2)\n")
#> 
#> 2. Benchmark: 'minkowski' distance (p=2)
minkowski_bench <- bench::mark(
  `distance(method = 'minkowski')` = distance(dist_matrix_rows, method = "minkowski", p = p, test.na = FALSE),
  `dist_many_many(method = 'minkowski')` = dist_many_many(dist_matrix_rows, dist_matrix_rows, method = "minkowski", p = p, testNA = FALSE),
  min_iterations = 3,  max_iterations = 20,
  check = \(a, b) all(a == b)
)
#> Metric: 'minkowski'; p = 2; comparing: 300 vectors.
#> Metric: 'minkowski'; p = 2; comparing: 300 vectors.
#> Metric: 'minkowski'; p = 2; comparing: 300 vectors.
#> Metric: 'minkowski'; p = 2; comparing: 300 vectors.
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
print(minkowski_bench)
#> # A tibble: 2 × 13
#>   expression      min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time
#>   <bch:expr>    <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm>
#> 1 distance(met… 270ms  309ms      3.31     217MB     6.62     3     6      906ms
#> 2 dist_many_ma… 314ms  330ms      3.03     284MB     8.09     3     8      989ms
#> # ℹ 4 more variables: result <list>, memory <list>, time <list>, gc <list>
plot(minkowski_bench)
```

![](https://i.imgur.com/2v454k3.png)<!-- -->

``` r

cat("\n3. Benchmark: 'jensen-shannon' distance (with unit)\n")
#> 
#> 3. Benchmark: 'jensen-shannon' distance (with unit)
jsd_bench <- bench::mark(
  `distance(method = 'jensen-shannon')` = distance(dist_matrix_rows, method = "jensen-shannon", unit = unit, test.na = FALSE),
  `dist_many_many(method = 'jensen-shannon')` = dist_many_many(dist_matrix_rows, dist_matrix_rows, method = "jensen-shannon", unit = unit, testNA = FALSE),
  min_iterations = 3, max_iterations = 20,
  check = \(a, b) all(a == b)
)
#> Metric: 'jensen-shannon' using unit: 'log2'; comparing: 300 vectors.
#> Metric: 'jensen-shannon' using unit: 'log2'; comparing: 300 vectors.
#> Metric: 'jensen-shannon' using unit: 'log2'; comparing: 300 vectors.
#> Metric: 'jensen-shannon' using unit: 'log2'; comparing: 300 vectors.
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
print(jsd_bench)
#> # A tibble: 2 × 13
#>   expression               min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
#>   <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>
#> 1 distance(method = … 791.59ms 808.94ms     1.21      217MB    2.02      3     5
#> 2 dist_many_many(met…    1.51s    1.51s     0.663     284MB    0.663     3     3
#> # ℹ 5 more variables: total_time <bch:tm>, result <list>, memory <list>,
#> #   time <list>, gc <list>
plot(jsd_bench)
```

![](https://i.imgur.com/rcCVAjr.png)<!-- -->


## After
``` r
# install.packages("bench")
library(philentropy)
library(bench)

set.seed(42)

n_dist <- 300 # Number of distributions
n_bins <- 200  # Number of bins per distribution

dist_matrix_rows <- matrix(runif(n_dist * n_bins), nrow = n_dist)

p <- 2
unit <- "log2"

cat("Benchmarking distance matrix functions with", n_dist, "distributions of", n_bins, "bins each.\n\n")
#> Benchmarking distance matrix functions with 300 distributions of 200 bins each.

cat("1. Benchmark: 'euclidean' distance\n")
#> 1. Benchmark: 'euclidean' distance
euclidean_bench <- bench::mark(
  `distance(method = 'euclidean')` = distance(dist_matrix_rows, method = "euclidean", test.na = FALSE),
  `dist_many_many(method = 'euclidean')` = dist_many_many(dist_matrix_rows, dist_matrix_rows, method = "euclidean", testNA = FALSE),
  min_iterations = 3,  max_iterations = 20,
  check = \(a, b) all(a == b)
)
print(euclidean_bench)
#> # A tibble: 2 × 13
#>   expression      min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time
#>   <bch:expr>   <bch:> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm>
#> 1 distance(me… 1.31ms  1.4ms      651.    1.92MB      0      20     0     30.7ms
#> 2 dist_many_m… 2.92ms 3.26ms      303.  706.63KB     16.0    19     1     62.6ms
#> # ℹ 4 more variables: result <list>, memory <list>, time <list>, gc <list>
plot(euclidean_bench)
```

![](https://i.imgur.com/a8PbYCQ.png)<!-- -->

``` r

cat("\n2. Benchmark: 'minkowski' distance (p=2)\n")
#> 
#> 2. Benchmark: 'minkowski' distance (p=2)
minkowski_bench <- bench::mark(
  `distance(method = 'minkowski')` = distance(dist_matrix_rows, method = "minkowski", p = p, test.na = FALSE),
  `dist_many_many(method = 'minkowski')` = dist_many_many(dist_matrix_rows, dist_matrix_rows, method = "minkowski", p = p, testNA = FALSE),
  min_iterations = 3,  max_iterations = 20,
  check = \(a, b) all(a == b)
)
print(minkowski_bench)
#> # A tibble: 2 × 13
#>   expression      min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time
#>   <bch:expr>   <bch:> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm>
#> 1 distance(me… 13.3ms 14.8ms      66.7    1.84MB     0       20     0      300ms
#> 2 dist_many_m… 34.2ms   36ms      27.8  703.17KB     2.14    13     1      468ms
#> # ℹ 4 more variables: result <list>, memory <list>, time <list>, gc <list>
plot(minkowski_bench)
```

![](https://i.imgur.com/XoBPrpn.png)<!-- -->

``` r

cat("\n3. Benchmark: 'jensen-shannon' distance (with unit)\n")
#> 
#> 3. Benchmark: 'jensen-shannon' distance (with unit)
jsd_bench <- bench::mark(
  `distance(method = 'jensen-shannon')` = distance(dist_matrix_rows, method = "jensen-shannon", unit = unit, test.na = FALSE),
  `dist_many_many(method = 'jensen-shannon')` = dist_many_many(dist_matrix_rows, dist_matrix_rows, method = "jensen-shannon", unit = unit, testNA = FALSE),
  min_iterations = 3, max_iterations = 20,
  check = \(a, b) all(a == b)
)
print(jsd_bench)
#> # A tibble: 2 × 13
#>   expression      min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time
#>   <bch:expr>   <bch:> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm>
#> 1 distance(me… 14.2ms 18.4ms      56.1    1.84MB     0       20     0      357ms
#> 2 dist_many_m… 39.3ms 42.4ms      23.5  703.17KB     2.14    11     1      467ms
#> # ℹ 4 more variables: result <list>, memory <list>, time <list>, gc <list>
plot(jsd_bench)
```

![](https://i.imgur.com/3YkfWgR.png)<!-- -->

</details>