@Nowosad 2025-10-01
Two more questions, @brownag :
1. do you have any example cases to check the performance (between the current and the proposed version)?
2. maybe a silly question, but is it possible to control the number of workers/cores used?