@brownag 2025-10-03
Hajk and Jakub,

Great, I am glad you approve. I will make the changes you asked for @HajkD and give it a thorough review again now knowing that you are supportive.

@Nowosad yes, that is a good summary of the PR. The templates are much more efficient and can be reused in high and low level functions via iterators as input. Also logic can be inlined into the loop for optimized compilation. They are written in pure C++ stdlib, with no calls to R API or other callbacks, so they have very little overhead comparatively and are thread-safe for parallelism.

For some example cases to check performance, see [motif::lsp_to_dist()](https://jakubnowosad.com/motif/reference/lsp_to_dist.html). It was when I was using this function a few months ago that I wondered whether I could speed up the philentropy calculations. I have a couple unrelated use cases involving soil profiles and distances in soil property/depth space that also can wind up with calculating many distances.

Consider this example based on the 5th vignette on clustering spatial patterns. I get about a 30-40x faster run time on my machine (16 cores). 

To test both in same reprex I have the current CRAN version of philentropy installed in my R library, then use `devtools::load_all()` to build from my workspace with this PR branch half way through.

To make it a bit more demanding, I decreased the window size from 100 to 50, so there are 4 times as many local landscapes as the vignette. It takes about 35 seconds to run with CRAN philentropy, and 1 second to run with this branch.

``` r
library(motif)
library(stars)
#> Loading required package: abind
#> Loading required package: sf
#> Linking to GEOS 3.12.1, GDAL 3.8.4, PROJ 9.4.0; sf_use_s2() is TRUE
library(sf)

safe_pal6 = c("#88CCEE", "#CC6677", "#DDCC77",
             "#117733", "#332288", "#888888")
safe_pal4 = c("#88CCEE", "#CC6677", "#DDCC77", "#117733")

landcover = read_stars(system.file("raster/landcover2015.tif", package = "motif"), proxy = TRUE)

landcover = droplevels(landcover)
plot(landcover, key.pos = 4, key.width = lcm(5), main = NULL)
#> downsample set to 8
```

![](https://i.imgur.com/l06gdZd.png)<!-- -->

``` r

landcover_cove = lsp_signature(landcover, type = "cove",
                               window = 50, normalization = "pdf")

resb1 <- system.time({
    landcover_dist = lsp_to_dist(landcover_cove, dist_fun = "jensen-shannon")
})
#> Metric: 'jensen-shannon' using unit: 'log2'; comparing: 4113 vectors.

print(resb1)
#>    user  system elapsed 
#>  40.708   0.203  40.926

resb2 <- system.time({
    landcover_dist = lsp_to_dist(landcover_cove, dist_fun = "jensen-shannon")
})
#> Metric: 'jensen-shannon' using unit: 'log2'; comparing: 4113 vectors.

print(resb2)
#>    user  system elapsed 
#>  34.518   0.173  34.695

resb3 <- system.time({
    landcover_dist = lsp_to_dist(landcover_cove, dist_fun = "jensen-shannon")
})
#> Metric: 'jensen-shannon' using unit: 'log2'; comparing: 4113 vectors.

print(resb3)
#>    user  system elapsed 
#>  35.684   0.194  35.882

devtools::load_all("~/workspace/philentropy")
#> ℹ Loading philentropy

resa1 <- system.time({
    landcover_dist = lsp_to_dist(landcover_cove, dist_fun = "jensen-shannon")
})

print(resa1)
#>    user  system elapsed 
#>   8.474   0.210   0.891

resa2 <- system.time({
    landcover_dist = lsp_to_dist(landcover_cove, dist_fun = "jensen-shannon")
})

print(resa2)
#>    user  system elapsed 
#>   9.083   0.220   1.050


resa3 <- system.time({
    landcover_dist = lsp_to_dist(landcover_cove, dist_fun = "jensen-shannon")
})

print(resa3)
#>    user  system elapsed 
#>   8.590   0.212   0.997

b = colMeans(rbind(resb1, resb2, resb3))
a = colMeans(rbind(resa1, resa2, resa3))

b/a
#>  user.self   sys.self    elapsed user.child  sys.child 
#>  4.2417868  0.8878505 37.9520082        NaN        NaN
```

I will fix the messages I disabled

2. RcppParallel has a function for number of threads, and also an environment variable.

```r
# Set number of threads before loading package
Sys.setenv("RCPP_PARALLEL_NUM_THREADS" = "4")
library(philentropy)

# Or use RcppParallel functions directly
RcppParallel::setThreadOptions(numThreads = 2)

# Check current setting
RcppParallel::defaultNumThreads()
```

This is a good thing to keep in mind as CRAN requires that tests and examples not use more than two threads. I can add `setThreadOptions(2)` to tests to handle this. Thanks for bringing that up.