@brownag 2025-10-09
@HajkD @Nowosad 

I have made a number of changes, all of which I hope to be improvements.

 - All (non-1:1) distance functions now accept a `num.threads` argument which is passed to the `parallelFor()` loop. If NULL (default) the RCPP_PARALLEL_NUM_WORKERS environment variable is consulted first and, in lieu of that being set, defaults to 2 threads. So, we can control the number of threads used on a per-call basis, and use the global environment variable if set. I think in general we do not want to default to what RcppParallel will choose for default (all available) as this might be unexpected for users who are accustomed to these routines running sequentially, or lead to issues especially in shared environments or where other jobs are running. For many use cases with modest input sizes, the overhead of setting up a dozen+ threads v.s. just 1 or 2 is probably not worthwhile.
 
 - The big if/else block, which was used in a couple places for dispatching the distance methods by name, which was moved from R to C++, has been abstracted into a pure C++ template that can be re-used in multiple places. This should make compiler and runtime optimization even better, simplifies the logic in the workers themselves, and keeps from having to make changes in multiple places if the logic needs changed

 - In the original set of changes, I was thinking we would need dedicated workers for the Minkowski-like functions. However, that setup didn't sit right with me, so I have consolidated the workers. Now the workers are less about the details of the distance methods themselves, but more about how the parallelism is implemented (i.e. over what input matrix), function signatures of the distance methods, and types of input objects. 
   - For minkowski, I was struggling a bit with the initialization of `p` as NA vs NULL, and proper handling downstream, so initially just separated these out. How we handle validation of that parameter and the logic around it is especially important in the C++ template functions which should not rely on Rcpp and be as simple as possible. I now have the input validation working before the parallel loop begins, so it should be efficient and generalizable.

 - The One-to-Many parallel worker has now been implemented to complement what was done with Many-to-Many. The parallelization strategy is different for this worker (i.e. parallel over Q vs. P). Because not all of the distance methods are symmetric I think we can't get away with just one (where P and Q are swapped)
 
 Please let me know if you have any questions, concerns, or requests for additional changes!