@HajkD 2025-10-31
@Nowosad Thank you for your swift reply!

Happy to merge now and push this to CRAN.

Cheers,
Hajk